% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptimalHbi.R
\name{optimalHbi}
\alias{optimalHbi}
\title{Variable selection using shrinkage priors}
\usage{
optimalHbi(bi, Hbi)
}
\arguments{
\item{bi}{The values of the tuning parameter}

\item{Hbi}{The estimated number of signals corresponding to each b.i}
}
\value{
the optimal value of tuning parameter and the associated H value
}
\description{
optimalHbi function will take b.i and H.b.i as input which comes from the result of TwoMeans function. It will return H: the optimal value of the tuning parameter.
}
\examples{
optimalHbi(bi, Hbi)
}
