% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S2MVarSelection.R
\name{S2MVarSelection}
\alias{S2MVarSelection}
\title{Variable selection using shrinkage priors:: Sequential2Means}
\usage{
S2MVarSelection(Beta, H)
}
\arguments{
\item{Beta}{matrix consisting of N posterior samples of p variables that is known either to user or from Sequential2Means function}

\item{H}{Estimated number of signals obtained from the optimal.b.i function}
}
\value{
Indices of important subset of variables
}
\description{
S2MVarSelection function will take S2M: a list obtained from the 2Means.variables function and H: the estimated number of signals obtained from the optimal.b.i function. This will give out the important subset of variables for the Gaussian Linear model.
}
\examples{

n <- 100
p <- 20
X <- matrix(rnorm(100), n, p)
beta <- exp(rnorm(p))
Y <- X \%*\% beta + rnorm(n, 0, 1)
df <- data.frame(X,Y)
rv.hs <- bayesreg::bayesreg(Y~. ,df, "gaussian", "horseshoe+", 10000, 2000)

Beta = rv.hs$beta
H = 10
impVariablesGLM = S2MVarSelection(Beta, H)
impVariablesGLM


}
