% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S2MVarSelection.R
\name{S2MVarSelectionV1}
\alias{S2MVarSelectionV1}
\title{Variable selection using shrinkage priors :: S2MVarSelectionV1}
\usage{
S2MVarSelectionV1(S2M, H = 10)
}
\arguments{
\item{S2M}{List obtained from the 2Means.variables function}

\item{H}{Estimated number of signals obtained from the optimal.b.i function}
}
\value{
Indices of important subset of variables for the Gaussian Linear model
}
\description{
S2MVarSelectionV1 function will take S2M: a list obtained from the 2Means.variables function and H: the estimated number of signals obtained from the optimal.b.i function. This will give out the important subset of variables for the Gaussian Linear model.
}
