% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2MeansVariables.R
\name{TwoMeans}
\alias{TwoMeans}
\title{Variable selection using shrinkage priors}
\usage{
TwoMeans(X, Y, t)
}
\arguments{
\item{X}{Design matrix}

\item{Y}{Response vector}

\item{t}{Vector of tuning parameter values from Sequential 2-means (S2M) variable selection algorithm.}
}
\value{
A list S2M which will hold p: the total number of variables, b.i: the values of the tuning parameter, H.b.i : the estimated number of signals corresponding to each b.i, abs.post.median: medians of the absolute values of the posterior samples.
}
\description{
2Means.variables function will take as input X: design matrix, Y : response vector, t: vector of tuning parameter values from Sequential 2-means (S2M) variable selection algorithm. The function will return a list S2M which will hold p: the total number of variables, b.i: the values of the tuning parameter, H.b.i : the estimated number of signals corresponding to each b.i, abs.post.median: medians of the absolute values of the posterior samples.
}
\examples{
TwoMeans(X, Y, t)
}
