% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sequential2Means.R
\name{Sequential2Means}
\alias{Sequential2Means}
\title{Variable selection using shrinkage priors}
\usage{
Sequential2Means(S2M, H)
}
\arguments{
\item{S2M}{List obtained from the 2Means.variables function}

\item{H}{Estimated number of signals obtained from the optimal.b.i function}
}
\value{
Important subset of variables for the Gaussian Linear model
}
\description{
Sequential2means function will take S2M: a list obtained from the 2Means.variables function and H: the estimated number of signals obtained from the optimal.b.i function. This will give out the important subset of variables for the Gaussian Linear model.
}
\examples{
Sequential2Means(S2M, H)
}
