% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sequential2Means.R
\name{Sequential2MeansBeta}
\alias{Sequential2MeansBeta}
\title{Variable selection using shrinkage prior:: Sequential2Means}
\usage{
Sequential2MeansBeta(Beta, lower, upper, l)
}
\arguments{
\item{Beta}{N by p matrix consisting of N posterior samples of p variables}

\item{lower}{the lower bound of the chosen values of the tuning parameter}

\item{upper}{the upper bound of the chosen values of the tuning parameter}

\item{l}{the number of chosen values of the tuning parameter}
}
\value{
A list S2M which will hold p: the total number of variables, b.i: the values of the tuning parameter, H.b.i : the estimated number of signals corresponding to each b.i, abs.post.median: medians of the absolute values of the posterior samples.
}
\description{
Sequential2MeansBeta function will take as input Beta : N by p matrix consisting of N posterior samples of p variables, lower : the lower bound of the chosen values of the tuning parameter, upper : the upper bound of the chosen values of the tuning parameter, and l :the number of chosen values of the tuning parameter. The function will return a list S2M which will hold p: the total number of variables, b.i: the values of the tuning parameter, H.b.i : the estimated number of signals corresponding to each b.i, abs.post.median: medians of the absolute values of the posterior samples.
}
\examples{



n <- 100
p <- 20
X <- matrix(rnorm(n*p), n, p)
beta <- exp(rnorm(p))
Y <- as.vector(X \%*\% beta + rnorm(n, 0, 1))
df <- data.frame(X,Y)
rv.hs <- bayesreg::bayesreg(Y~. ,df, "gaussian", "horseshoe+", 200, 100)

Beta = t(rv.hs$beta)
lower = 0
upper = 1
l = 20
S2Mbeta = Sequential2MeansBeta( Beta, lower, upper, l )
H.b.i = S2Mbeta$H.b.i


#' n <- 100
p <- 20
X <- matrix(rnorm(n*p), n, p)
beta <- exp(rnorm(p))
Y <- as.vector(X \%*\% beta + rnorm(n, 0, 1))
df <- data.frame(X,Y)
rv.hs <- bayesreg::bayesreg(Y~. ,df, "normal", "lasso", 200, 100)

Beta = t(rv.hs$beta)
lower = 0
upper = 1
l = 15
S2Mbeta = Sequential2MeansBeta( Beta, lower, upper, l )
H.b.i = S2Mbeta$H.b.i

}
